requirements.txt
requests
psycopg2-binary
prometheus_client